+++ src/views/components/Pagination/index.jsx
@@ -3,13 +3,13 @@ import Pagination from '@mui/material/Pagination';
 import Stack from '@mui/material/Stack';
 
 const Paginations = () => {
-    return (
-        <div className='mt-10 flex justify-end'>
-            <Stack spacing={2}>
-                <Pagination count={10} variant="outlined" shape="rounded" />
-            </Stack>
-        </div>
-    )
+  return (
+    <div className='mt-10 flex justify-end'>
+      <Stack spacing={2}>
+        <Pagination count={10} variant="outlined" shape="rounded" />
+      </Stack>
+    </div>
+  )
 }
 
 export default Paginations
