+++ src/views/components/SearchField/index.jsx
@@ -1,26 +1,25 @@
 import { FormControl, InputAdornment, TextField } from "@mui/material";
-import React, { useState } from "react";
+import React from "react";
 import SearchIcon from '@mui/icons-material/Search';
 
 const SearchField = () => {
+  return (
+    <div>
+      <FormControl className='m-0'>
+        <TextField
+          size="small"
+          variant="outlined"
+          InputProps={{
+            startAdornment: (
+              <InputAdornment position="start">
+                <SearchIcon />
+              </InputAdornment>
+            )
+          }}
+        />
+      </FormControl>
+    </div>
+  )
+};
 
-    return (
-        <div>
-            <FormControl className='m-0'>
-                <TextField
-                    size="small"
-                    variant="outlined"
-                    InputProps={{
-                        startAdornment: (
-                            <InputAdornment position="start">
-                                <SearchIcon />
-                            </InputAdornment>
-                        )
-                    }}
-                />
-            </FormControl>
-        </div>
-    )
-}
-
-export default SearchField
+export default SearchField;
